create kotban-style kanban board from existing dictionary
since the development of IPA-Transcribers
  to produce more phonologically accurate transcriptions of native words,
and changes in Tuki Nuwa's phonology,

Large portions of the current dictionary are now out of date.

Out of date in the sense that, either the lexeme chosen is either:
* no longer a legal word, given the new phonotactics
* no longer the best choice for international recognisability,
  given the more accurate comparisons between world languages

While working, I've also found simple text files with one per lexeme
to be too limiting.

Given the large number of states a word can now be in, in terms of finishedness,
I knew I needed a more complex workflow to handle new word ideas,
and reprocess existing ones.

I chose kanban, but wanted changes in the kanban board to be storable in git;
I couldn't find a suitable existing kanban viewer, so I wrote my own.

The changes in THIS commit create the initial board,
but only from the old dictionary;
later commits need to add words (and possible words) from the other text files.