w is dead, long live v
I noticed that if you count f, v and w as one letter,
you get a good propertion of world speakers
who are familiar with at least one of those sounds.

Considering we already combine sounds that are distinct in English into
one letter in Tokaja (l and r),
it's self-consistent to use the letter v to represent sounds close to f, v or w.

So the letter v can be pronounced as w, f or v